requirements.txt
python-dotenv==1.1.1
requests==2.32.4
paramiko==4.0.0
